+++ tests/Sabre/XML/ReaderTest.php
@@ -232,6 +232,48 @@ BLA;
 
     }
 
+    /**
+     * @depends testMappedElementCallBack
+     */
+    function testReadText() {
+
+        $input = <<<BLA
+<?xml version="1.0"?>
+<root xmlns="http://sabredav.org/ns">
+  <elem1>
+    <elem2>hello </elem2>
+    <elem2>world</elem2>
+  </elem1>
+</root>
+BLA;
+
+        $reader = new Reader();
+        $reader->elementMap = [
+            '{http://sabredav.org/ns}elem1' => function(Reader $reader) {
+                return $reader->readText();
+            }
+        ];
+        $reader->xml($input);
+
+        $output = $reader->parse();
+
+        $expected = [
+            'name' => '{http://sabredav.org/ns}root',
+            'value' => [
+                [
+                    'name' => '{http://sabredav.org/ns}elem1',
+                    'value' => 'hello world',
+                    'attributes' => [],
+                ],
+            ],
+            'attributes' => [],
+
+        ];
+
+        $this->assertEquals($expected, $output);
+
+    }
+
     function testParseProblem() {
 
         $input = <<<BLA