+++ ChangeLog.md
@@ -4,8 +4,11 @@ ChangeLog
 0.1.0 (????-??-??)
 ------------------
 
+* #16: Added ability to override `elementMap`, `namespaceMap` and `baseUri` for
+  a fragment of a document during reading an writing using `pushContext` and
+  `popContext`.
 * Removed: `Writer::$context` and `Reader::$context`.
-* Added: `Reader::$baseUri` to match `Writer::$baseUri`.
+* #15: Added `Reader::$baseUri` to match `Writer::$baseUri`.
 
 
 0.0.6 (2014-09-26)